+++ engine/flink/management/periodic/src/test/scala/pl/touk/nussknacker/engine/management/periodic/PeriodicProcessGen.scala
@@ -1,11 +1,30 @@
 package pl.touk.nussknacker.engine.management.periodic
 
+import pl.touk.nussknacker.engine.api.ProcessVersion
 import pl.touk.nussknacker.engine.build.EspProcessBuilder
 import pl.touk.nussknacker.engine.canonicalgraph.CanonicalProcess
 import pl.touk.nussknacker.engine.management.periodic.CronSchedulePropertyExtractor.CronPropertyDefaultName
+import pl.touk.nussknacker.engine.management.periodic.model.{DeploymentWithJarData, PeriodicProcess, PeriodicProcessId}
+
+import java.time.LocalDateTime
 
 object PeriodicProcessGen {
-  def apply(cronProperty: String = "0 0 * * * ?"): CanonicalProcess = {
+  def apply(): PeriodicProcess = {
+    PeriodicProcess(
+      id = PeriodicProcessId(42),
+      deploymentData = DeploymentWithJarData(
+        processVersion = ProcessVersion.empty,
+        canonicalProcess = buildCanonicalProcess(),
+        inputConfigDuringExecutionJson = "{}",
+        jarFileName = "jar-file-name.jar"
+      ),
+      scheduleProperty = CronScheduleProperty("0 0 * * * ?"),
+      active = true,
+      createdAt = LocalDateTime.now(),
+    )
+  }
+
+  def buildCanonicalProcess(cronProperty: String = "0 0 * * * ?"): CanonicalProcess = {
       EspProcessBuilder
         .id("test")
         .additionalFields(properties = Map(CronPropertyDefaultName -> cronProperty))